docs(README): fix link to API documentation (#83)
Previous link was a 404. I suppose you could either link to the main project page or the docs, but given that the main product page is linked to already as the "Product Documentation", I assume the intent for this link is to go to the docs themselves.

This is a trivial change so the template checks don't really apply.